Change to using jquery initialize plugin
Previous solution didn't account for DataTable being reloaded WITHOUT the page being refreshed when switching between apps, so only on the first page load would the script run.

New solution works in this case and doesn't show the toolbar for a split second before hiding it. All around better.

Initialize plugin works only in IE9+ (please let this not be an issue).